requirements/common.txt
Django==1.7.6
django-admin-honeypot==0.4.0
django-model-utils==2.2
django-mptt==0.6.1
django-pipeline==1.4.3
django-redis==3.5.1
django-solo==1.1.0
django-suit==0.2.12
django-suit-ckeditor==0.0.2
beautifulsoup4==4.3.2
Pillow==2.7.0
pysyge==0.2.0
pytils==0.3
psycopg2==2.6
redis==2.10.3
requests==2.5.3
sqlparse==0.1.14